requirements.txt
# Core dependencies
openai>=1.0.0
google-generativeai>=0.3.0
tweepy>=4.14.0
Pillow>=10.0.0
schedule>=1.2.0
requests>=2.31.0
python-dotenv>=1.0.0

# Optional dependencies for enhanced functionality
python-crontab>=2.7.0
APScheduler>=3.10.0

# Development dependencies (uncomment if needed)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0